Fix request for incorrect nodelist
Apart from being no longer accesible on port 8443, the startnodes.lst
file is not always fully compatible with the Windows Miner, and
fullnodes.lst should be used instead.

The fullnodes.lst file is much like startnodes.lst, except the list is
filtered by nodes that allow a pull of the latest candidate block from
a Linux node, which is a critical dependency of the Windows Miner.
Additionally, the use of port 8443 is no longer required to access the
list. This removes a limitation for people in certain areas where ports
may be restricted. The port will remain open as legacy support until a
point where it is provably no longer in use, after which communication
on that port will be disabled.