Fix nodelist download/update issue
There is an issue with the Coreplist not updating correctly.
I think the issue was due to separate compilation of the C++ files
which resulted in each file holding it's own copy of the Coreplist
due to the static assignment. Therefore an update from a function in
one file does not represent an update to the same Coreplist in another
file and, as a result, the Coreplist is never truly updated where it
matters.

The fix I implemented INCLUDES the required files into the winminer.cpp
file which allows all functions to use the same static Coreplist under
the same scope. This results in an easier compilation process using the
Visual Studio Developer Command Prompt for anyone wishing to compile
their own binaries.

I've also included some VERY BASIC instructions on how to compile.