Block update improvements
Miners with a good connection to the network are grabbing candidate
blocks with only 1 transaction in them and solving that block until the
next block update. This will be problematic to the network in the
not-too-distant future when transactions on the network are sent
throughout the block. Unfortunately there is currently no way to
acquire the lastest transaction count from a node. So, I've implemented
an intra-block update mechanism to the update-monitor which forces the
miner to download a new candidate block at 40, 120, 280 & 600 seconds
after the first block update has been detected. It's an exponentially
increasing update interval so as to reduce bandwidth of the miner
downloading multiple candidate blocks, but still captures most of the
early block transactions.